refactor: 패키지구조 재설정
