+++ src/main/java/com/github/chacha89/todos/user/controller/Controller.java
@@ -1,0 +1,4 @@
+package com.github.chacha89.todos.user.controller;
+
+public class Controller {
+}
