+++ src/main/java/com/github/chacha89/todos/user/dto/requestDto/APIRequest.java
@@ -1,0 +1,4 @@
+package com.github.chacha89.todos.user.dto.requestDto;
+
+public class APIRequest {
+}
