+++ src/main/java/com/github/chacha89/todos/user/dto/responseDto/APIResponse.java
@@ -1,0 +1,4 @@
+package com.github.chacha89.todos.user.dto.responseDto;
+
+public class APIResponse {
+}
