+++ src/main/java/com/github/chacha89/todos/user/exception/GlobalExceptionHandler.java
@@ -1,0 +1,4 @@
+package com.github.chacha89.todos.user.exception;
+
+public class GlobalExceptionHandler {
+}
