+++ src/main/java/com/github/chacha89/todos/user/repository/Repository.java
@@ -1,0 +1,7 @@
+package com.github.chacha89.todos.user.repository;
+
+import com.github.chacha89.todos.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface Repository extends JpaRepository<User,Long> {
+}
