+++ src/main/java/com/github/chacha89/todos/user/service/Service.java
@@ -1,0 +1,4 @@
+package com.github.chacha89.todos.user.service;
+
+public class Service {
+}
